endorser transaction: compute envelope
